+++ src/storm/storage/dd/sylvan/InternalSylvanAdd.h
@@ -612,6 +612,16 @@ namespace storm {
              */
             static ValueType getValue(MTBDD const& node);
 
+            /*!
+             * Retrieves whether the topmost variable in the MTBDD is the one with the given index.
+             *
+             * @param The top node of the MTBDD.
+             * @param variableIndex The variable index.
+             * @param offset An offset that is applied to the index of the top variable in the MTBDD.
+             * @return True iff the MTBDD's top variable has the given index.
+             */
+            static bool matchesVariableIndex(MTBDD const& node, uint64_t variableIndex, int64_t offset = 0);
+            
         private:
             /*!
              * Recursively builds the ODD from an ADD.