+++ src/storm/storage/dd/sylvan/InternalSylvanBdd.cpp
@@ -283,7 +283,7 @@ namespace storm {
             // If we are at the maximal level, the value to be set is stored as a constant in the DD.
             if (currentRowLevel == maxLevel) {
                 result.set(currentRowOffset, true);
-            } else if (ddRowVariableIndices[currentRowLevel] < sylvan_var(dd)) {
+            } else if (bdd_isterminal(dd) || ddRowVariableIndices[currentRowLevel] < sylvan_var(dd)) {
                 toVectorRec(dd, result, rowOdd.getElseSuccessor(), complement, currentRowLevel + 1, maxLevel, currentRowOffset, ddRowVariableIndices);
                 toVectorRec(dd, result, rowOdd.getThenSuccessor(), complement, currentRowLevel + 1, maxLevel, currentRowOffset + rowOdd.getElseOffset(), ddRowVariableIndices);
             } else {
@@ -390,7 +390,7 @@ namespace storm {
             
             if (currentLevel == maxLevel) {
                 result[currentIndex++] = values[currentOffset];
-            } else if (ddVariableIndices[currentLevel] < sylvan_var(dd)) {
+            } else if (bdd_isterminal(dd) || ddVariableIndices[currentLevel] < sylvan_var(dd)) {
                 // If we skipped a level, we need to enumerate the explicit entries for the case in which the bit is set
                 // and for the one in which it is not set.
                 filterExplicitVectorRec(dd, currentLevel + 1, complement, maxLevel, ddVariableIndices, currentOffset, odd.getElseSuccessor(), result, currentIndex, values);
@@ -424,7 +424,7 @@ namespace storm {
             
             if (currentLevel == maxLevel) {
                 result.set(currentIndex++, values.get(currentOffset));
-            } else if (ddVariableIndices[currentLevel] < sylvan_var(dd)) {
+            } else if (bdd_isterminal(dd) || ddVariableIndices[currentLevel] < sylvan_var(dd)) {
                 // If we skipped a level, we need to enumerate the explicit entries for the case in which the bit is set
                 // and for the one in which it is not set.
                 filterExplicitVectorRec(dd, currentLevel + 1, complement, maxLevel, ddVariableIndices, currentOffset, odd.getElseSuccessor(), result, currentIndex, values);
@@ -525,6 +525,10 @@ namespace storm {
             
             // Return the variable for this node.
             return newNodeVariable;
+        }
+        
+        bool InternalBdd<DdType::Sylvan>::matchesVariableIndex(BDD const& node, uint64_t variableIndex, int64_t offset) {
+            return !sylvan_isconst(node) && static_cast<uint64_t>(sylvan_var(node) + offset) == variableIndex;
         }
         
         template InternalAdd<DdType::Sylvan, double> InternalBdd<DdType::Sylvan>::toAdd() const;